requirement.txt
absl-py==1.4.0
altgraph==0.17.3
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astunparse==1.6.3
awscli==1.27.135
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.26.4
black==21.4b2
boto3==1.26.135
botocore==1.29.135
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.4
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@82a57ce0b70057685962b352535147d9a8118578
docutils==0.16
filelock==3.12.0
flatbuffers==23.3.3
fonttools==4.39.4
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.4.0
google-api-core==2.11.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-cloud-vision==3.4.1
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
h5py==3.8.0
hydra-core==1.3.2
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
install==1.3.5
iopath==0.1.9
isodate==0.6.1
jax==0.4.8
Jinja2==3.1.2
jmespath==1.0.1
keras==2.12.0
kiwisolver==1.4.4
layoutparser==0.3.4
lazy_loader==0.2
libclang==16.0.0
macholib==1.16.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mpmath==1.3.0
msrest==0.7.1
mypy-extensions==1.0.0
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
portalocker==2.7.0
proto-plus==1.22.2
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser==2.21
pydot==1.4.2
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==5.4.1
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.6.1
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
tifffile==2023.4.12
toml==0.10.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Wand==0.6.11
Werkzeug==2.3.4
wrapt==1.14.1
yacs==0.1.8
zipp==3.15.0
